requirements.txt
typer
PyYAML
Scrapy
setuptools
SQLAlchemy
pandas
Jinja2
itemadapter
rich
pymysql
requests
pydantic